# travel-api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Travel Api",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/rental-companies/{rentalCompanyId}": {
      "get": {
        "description": "- URL\n    - /rental-companies/\n- Method\n    - GET\n- Response\n    \n\n<img src=\"https://content.pstmn.io/def81b3d-2a2f-4cd7-b184-e3d82839bb01/aW1hZ2UucG5n\" alt=\"\" height=\"629\" width=\"682\">",
        "summary": "Get rental company by ID",
        "tags": [
          "Rental Companies"
        ],
        "operationId": "GetrentalcompanybyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "rentalCompanyId",
            "in": "path",
            "description": "ID of the Rental Company",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "- URL\n    - /rental-companies/\n- Method\n    - GET\n- Response\n  ",
        "summary": "Edit rental company by ID",
        "tags": [
          "Rental Companies"
        ],
        "operationId": "PutrentalcompanybyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "rentalCompanyId",
            "in": "path",
            "description": "ID of the Rental Company",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/EditRentalCompanyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "- URL\n    - /rental-companies\n- Method\n    - POST\n- Headers\n    - x-access-token : Bearer\n- Response",
        "summary": "Delete rental company by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Rental Companies"
        ],
        "operationId": "DeleterentalcompanybyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "rentalCompanyId",
            "in": "path",
            "description": "ID of the Rental Company",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/cars/{carId}": {
      "get": {
        "description": "- URL\n    - /cars/\n- Method\n    - GET\n- Response\n    \n\n<img src=\"https://content.pstmn.io/1a46bd67-565d-4f08-b853-ac2b570a135f/aW1hZ2UucG5n\" alt=\"\" height=\"335\" width=\"819\">",
        "summary": "Get car by ID",
        "tags": [
          "cars"
        ],
        "operationId": "GetcarbyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "description": "ID of the Car",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "- URL\n    - /cars\n- Method\n    - POST\n- Headers\n    - x-access-token : Bearer\n- Request Body\n    - **name** as **string**\n    - **type** as **string**\n    - **year** as **string**\n    - **rental_company_id as string**\n- Response\n    \n\n",
        "summary": "Edit car with valid payload and admin role",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "cars"
        ],
        "operationId": "Editcarwithvalidpayloadandadminrole",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/EditCarRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "- URL\n    - /cars/\n- Method\n    - DELETE\n- Headers\n    - x-access-token : Bearer\n- Response\n    \n\n<img src=\"https://content.pstmn.io/729218ed-df44-45d3-aa2e-da2b5120650f/aW1hZ2UucG5n\" alt=\"\" height=\"137\" width=\"741\">",
        "summary": "Delete car by id",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "cars"
        ],
        "operationId": "Deletecarbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "description": "ID of the Car",
            "required": true,
            "type": "string"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/cars/{carId}/prices": {
      "post": {
        "description": "- URL\n    - /cars//prices\n- Method\n    - POST\n- Headers\n    - x-access-token : Bearer\n- Request Body\n    - **rental_price** as **integer**\n- Response\n    \n\n<img src=\"https://content.pstmn.io/4c9480bb-3e29-4954-85d3-37900044cb7b/aW1hZ2UucG5n\" width=\"805\" height=\"290\">",
        "summary": "Add rental price",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Rental Price"
        ],
        "operationId": "Addrentalprice",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "priceId",
            "in": "path",
            "description": "ID of the Price",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddrentalpriceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/cars/{carId}/prices/{priceId}": {
      "delete": {
        "summary": "Delete rental price by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Rental Price"
        ],
        "operationId": "DeleterentalpricebyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "priceId",
            "in": "path",
            "description": "ID of the Price",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/locations/{locationId}": {
      "get": {
        "summary": "Get Location By ID",
        "tags": [
          "Locations"
        ],
        "operationId": "GetLocationByID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "ID of the Location",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Edit Location With Valid Payload",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Locations"
        ],
        "operationId": "EditLocationWithValidPayload",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": false,
            "enum": [
              "application/json"
            ],
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/EditLocationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete Location By ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Locations"
        ],
        "operationId": "DeleteLocationByID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "ID of the Location",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    }
  },
  "definitions": {
...
    "EditRentalCompanyRequest": {
      "title": "EditRentalCompanyRequest",
      "example": {
        "name": "Rental Company A",
        "address": "jalan Samping",
        "telephone": "0822-xxxx-xxxx"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "telephone": {
          "type": "string"
        }
      }
    },
...
    "EditCarRequest": {
      "title": "EditcarwithvalidpayloadandadminroleRequest",
      "example": {
        "rental_company_id": "rental_company-y1zLnmYomr",
        "name": "Mitsubishi Pajero Sport",
        "type": "SUV",
        "year": "2022-03-03",
        "status": "AVAILABLE"
      },
      "type": "object",
      "properties": {
        "rental_company_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "year": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "required": [
        "rental_company_id",
        "name",
        "type",
        "year",
        "status"
      ]
    },
    "AddrentalpriceRequest": {
      "title": "AddrentalpriceRequest",
      "example": {
        "rental_price": 1200000
      },
      "type": "object",
      "properties": {
        "rental_price": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "rental_price"
      ]
    },
...
    "EditLocationRequest": {
      "title": "EditLocationWithValidPayloadRequest",
      "example": {
        "rental_company_id": "rental_company-AYc0qbgvdC",
        "name": "Location C",
        "address": "Your Location"
      },
      "type": "object",
      "properties": {
        "rental_company_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        }
      },
      "required": [
        "rental_company_id",
        "name",
        "address"
      ]
    }
  },
...
}